security update

+++ reset.php
@@ -14,26 +14,31 @@
     <?php
       include "utilities.php";
       //echo $admin;
-      $link=connect();
-      $dotaz="DROP TABLE $db_table";
-      if(mysqli_query($link, $dotaz)){debug ("OK - Tabulka $db_table byla úspěšně smazána");}
-      else{
-        debug("NOTICE - tabulka $db_table nebyla úspěšně smazána pravděpodobně proto, že ještě neexistuje $dotaz");
-      }
-      $dotaz="CREATE TABLE $db_table (
-        Id INT UNIQUE PRIMARY KEY,
-        Name TEXT,
-        Zasuvka INT,
-        Wifina INT,
-        LatLon TEXT,
-        Confirms INT,
-        Reports INT,
-        Popis TEXT,
-        Obrazek TEXT
-      );";
-      if(mysqli_query($link, $dotaz)){echo "OK - Tabulka $db_table byla úspěšně vytvořena";}
-      else{
-        echo "ERROR - tabulka $db_table nebyla vytvořena";
+      if($_REQUEST["username"]==$admin && $_REQUEST["password"]==$admin_password){
+        $link=connect();
+        $dotaz="DROP TABLE $db_table";
+        if(mysqli_query($link, $dotaz)){debug ("OK - Tabulka $db_table byla úspěšně smazána");}
+        else{
+          debug("NOTICE - tabulka $db_table nebyla úspěšně smazána pravděpodobně proto, že ještě neexistuje $dotaz");
+        }
+        $dotaz="CREATE TABLE $db_table (
+          Id INT UNIQUE PRIMARY KEY,
+          Name TEXT,
+          Zasuvka INT,
+          Wifina INT,
+          LatLon TEXT,
+          Confirms INT,
+          Reports INT,
+          Popis TEXT,
+          Obrazek TEXT
+        );";
+        if(mysqli_query($link, $dotaz)){echo "OK - Tabulka $db_table byla úspěšně vytvořena";}
+        else{
+          echo "ERROR - tabulka $db_table nebyla vytvořena";
+        }
+        else {
+          echo "ERROR - Login failed";
+        }
       }
     ?>
   </body>
